requirements.txt
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
python-telegram-bot-calendar==1.0.5
requests==2.28.1
aiohttp==3.8.3
